starting llamaguard finetuning

+++ white_box/monitor.py
@@ -79,10 +79,10 @@ class ActMonitor():
         self.device = device
 
     def predict_proba(self, acts):
-        return self.probe.predict_proba(acts)
+        return self.probe.predict_proba(acts.float())
 
     def get_loss(self, acts : torch.tensor):
         if len(self.tok_idxs) > 1:
-            return self.predict_proba(acts).mean(dim = -1)
+            return self.predict_proba(acts.float()).mean(dim = -1)
         else:
-            return self.predict_proba(acts)
+            return self.predict_proba(acts.float())
